query input events one by one
The way the pollEvent() worked before was to retrieve all available messages,
but only return the very last one to the ScummVM system.

The way it is supposed to work is to return all of them one by one,
so ScummVM can dispatch them individually.

I noticed that sometimes mousecklicks did not seem to register and I suspect
this could be the fix for that issue.