fix mixer thread exit
When not playing a sound, the mixer thread loop would not
check for SIGBREAKF_CTRL_C. After a delay of 10 frames, the main threasd would start
tearing down the mixer, leaving the mixerthread access an now-invalid g_mixer.

The new method implements a handshake, so the main thread would not blindly
tear down everything before the mixer thread is actually finished.