MidiDriver_CAMD: protect access to midi link via mutex
It seems, it may happen that both the main thread and the MIDI (timer)
thread try to send MIDI commands at the same time.

Happens in MonkeyIsland II intro sequence when MIDI only + MT32 is selected.
Crashed the whole system as it apparently in the CAMD's serial transmit
interrupt function(?)