Fix crash on second startup.
I used to get a lot of crashes or enforcer hits afetr starting ScummVM
for the second time.
Turns out, '_hardwareScreenBuffer[]' was not initialized to 0 every time
and the way the graphics setup works in endGFXTransaction() is to call
unloadGFXMode() before loadGFXMode() has even being called a first time.
Thus, we would call FreeScreenBuffer() on random adresses. It probably
worked the first time when all memory was zeroed.

Fix  by initializing _hardwareScreenBuffer[] to zero in the ctor (as we
should have done right from the start :-/ )