Remove need for virtual inheritance and dynamic_cast in SCI
There's really no need to downcast, if you're always require the downcast to work.
In particular makeAIFFStream returns a RewindableAudioStream but in most cases
the calling code requires the returned stream to be of type SeekableAudioStream.
Only two rare types of streams were actually not of type SeekableAudioStream in first
place (ADP4 and SDX2), so special-case the code for just the two.