Don't automatically close PopupDialog after 300ms
The existing code would close the PopupDialog if the first mouseup event
arrives more than 300ms after opening the dialog. On Amiga opening the dialogs
takes an aweful lot of time, so it in 'scummmodern' theme the popus would always
close immediately after clicking on them once (you'd need to hold the button to
leave the popup open).

For now, remove the 300ms time constraint.