sci: limit searching volume files past 2kb
For some reason, with some optimization settings, the loop would
almost never finish, or very slowly... as if feof() did not work properly.

The old code would alternatively try to search up to 1mb into the file.
This obviously takes a long time... which may seem as if the program is hung.

Try to work around this by limiting the search to the first 2kb.