sci: specialize 320pixel wide pixel drawing routines
The 320x200/320x240 modes are the ones we're most interested in to
accelerate on the Amiga.
Instead of making runtime decisions for upscaling and a dynamic width,
provide a specialized putPixel function with a hardcoded size of 320.
Similarly, implement a specialized getPixel320() function.

An alternative could be to use an abstract 'PutPixel' class and provide
specialized implementations - this boils down to function pointers (vtable) again.
I'm also thinking of using template specialization for 'drawMask'

Also const-ify functions in the hope to help the compiler a bit.
In Addition, use __regargs for the functions - albeit it doesn't seem the
compiler will adhere to that.

Use regparm(3), which seems to be optimal for the putPixel function