sci: Use an Array instead of HashMap for all object storage
This allows muchfaster access to the objects via just their offset -
instead of going through a hashmap, we directly use the offset
to access an array of object pointers. In particular one of the high-hitting
functions SegManager::getObject() will benefit from that.

This is only enabled, when SCI32 is disabled, since SCI32 allows much
bigger offsets. Not sure if we can use the array approach there, too
without having to create a really huge array. I did not investigate into
the sparseness of the object IDs (offsets).

One wart in this approach is places where we need to traverse the list
of live objects. There can be potentially unused slots, so we have to skip
them when traversing the array. Hide this ugliness behind a macro.