amigaos3: fix end-of-file handling
The original implementation signaled eof as soon as one reads to the end
of the file, but  feof() is required to signal only after one attempts to read at or past _beyond_
the end of file.

"This indicator is generally set by a previous operation on the stream that attempted to read
at or past the end-of-file.
Notice that stream's internal position indicator may point to the end-of-file for the next
operation, but still, the end-of-file indicator may not be set until an operation attempts
to read at that point.
This indicator is cleared by a call to clearerr, rewind, fseek, fsetpos or freopen.
Although if the position indicator is not repositioned by such a call, the next i/o
operation is likely to set the indicator again."