SCI: speed up Object::locateVarSelector
locateVarSelector() is one of the most often called functions.
It tries to find the variable index for a given selector.
It used to use a linear search through _baseVars to find the
variable carrying the given selector.

Now we're sorting the _baseVars by selector and store the variable index
alongside the selector. This is needed, since the sorting of _baseVars,
its elements shift around. So there's no longer a 1:1 mapping between
_baseVars and _variables.

This allows to use bisection to search for selector Ids.

Overall, this roughly doubles the lookup in speed.