Remove the need for RTTI
When calling DefaultSaveFileManager::concatWithSavesPath we can pretty much
assume that the SaveFileManager singleton is a DefaultSaveFileManager.
At least the places where concatWithSavesPath is actually used, make
sure of that. And those places are not being compiled on AmigaOS anyway.

DumpFile::seek() and size() seem to only work when the internal
filehandle is of type SeekableWriteStream. On AmigaOS, the
AmigaOSFile class is only of type WriteStream.