amigaos: remove pthread dependency
Apprently there are a few static objects in the build and by default
the creation of those is made threadsafe. With the introduction of pthread
to the amigaos toolchain, this would have required to link against
pthread.
We don't actually have multiple threads sharing static objects, so just
disable thread safety with -fno-threadsafe-statics