AMIGAOS3: Add optional default timer/AGA backend usage, fix FSNode::getParent
Adds the option to force the AGA backend and/or use the default timer, supposed to
be controlled using tooltypes, but these currently don't seem to work at all.
Also fixes getParent so that it actually returns the drawer selected rather than
the name of the first file found in said drawer.