AMIGAOS3: Use a per-Task TimerIOReq
OSystem::delayMillis may get called from multiple threads,
but the TimerIOReq struct in use can only be used from one at a time.

To get around this, introduce a 'task local' timer which is automatically
associated with the tasks in play.

On the Amiga there can be potentially 3 Task executing code in parallel:
the main thread, the timer thread and the mixer thread.