GCC's linker doesn't like inine virtual functions
The issue is that if the user of a base class doesn't know about
or include potential sublcasses. This way one compilation unit may never
encounter the implementation of a subclass' virtual function and then
the linker will not know where to find the actual implementation.

Since virtual functions can't be inlined anyway; prevent this scenario by moving
their implementation out of the header into a .cpp file.